Feature/chitin sync (#2)
* Remove Cargo.lock from source control

The Cargo.lock file is typically not included in version-controlled repositories for libraries, as it locks dependencies for the library's consumers. By removing it, we ensure that downstream users resolve dependencies based on their own requirements and constraints.

* fixed chitin view by creating new view to get around super strange rendering issues,

changes the vuetify tabs to ant tabs, also fixing rendering problems

* Add ChitinView component for displaying Chitin.key data

This new Vue component implements a detailed UI to visualize metadata, resource keys, filenames, and file entries from Chitin.key files. It includes tables, statistics, and tabs, and integrates AuroraService to load and process the data dynamically.